Split views.py into separate modules
